requirements.txt
faiss-cpu==1.7.4
sentence-transformers==2.7.0
torch==2.2.2
numpy==1.26.4
altair==4.2.2
google-generativeai==0.5.2
flask==2.3.3
line-bot-sdk==3.5.0
requests==2.31.0
gunicorn==21.2.0